fix: docs examples python version in tests
When performing test on docs example use the version of Python that is
specified by the test block to set the target version of Python.

This fixes failures with later ruff versions which validate the syntax
according to the target Python version more strictly, which was
resulting in a failure then except* was used in the example code against
Python 3.9.