chore: skip Gemini tests in local runs
Skip Gemini example tests in local runs to prevent it failing when the
user has gcloud configured but not with the ability to run the Gemini
APIs as that prevents coverage profiles from being run when developing
new features or fixing bugs.

Switch test skip from a print to pytest.skip so that its clear that the
test didn't actually run.

Fixes: #2167