FunctionTextFormat -> TextFormat, change handling
This separates the implementation of regex and lark to two classes.
The base class of these would be empty, so the `TextFormat` has become a
union type.
The openai tool call handling has been changed to silently ignore
formats that it does not handle. This is consistent with how the gpt-5
models ignore temperature parameter, which is not supported by reasoning
models (see https://github.com/pydantic/pydantic-ai/pull/2483).